[http] Refactor: Breaking the MVC and avoiding confusion with k8s ctrls
* Renamed controllers to http
* A more clean arch / onion / hexagonal approach
