+++ pkg/http/handlers/api_handler.go
@@ -1,4 +1,4 @@
-package api
+package handlers
 
 import (
 	"fmt"
